requirements.txt
tensorflow
tensorflow-datasets
cleverhans

absl-py # may install with tf
easydict # may install with tf